Refresh display iframes for signage every 60 seconds.
The display iframes are the only pieces of content submitted to their
respective content fields on Concerto, and have historically been refreshed by
Concerto every two minutes.  This appears to have stopped working reliably on
or about 2019-04-26.

https://github.com/concerto/concerto/wiki/Field-Configuration/d9481e7212c9d0f29bd90c0f77efbe313011534d
mentions that fields can be configured with `repeat_content`, which "specifies
whether or not the frontend should attempt to reload content if it is the same
as what is currently displayed. The default value, if not specified, is
'Suppress'.  To enable the frontend to reload the same piece of content (as
opposed to just leaving it on the screen for another duration) set this value
to 'Reload'."

So it's unclear how this was working in the first place.  Add this meta-refresh
in addition to setting `repeat_content` on the server in order to have the best
chances of success, so we're not solely reliant on the meta-refresh succeeding
either.