Show exact times for bookings on day/week views.
Don't just limit this to month view, as new scheduling procedures effective
Fall 2021 will mean class times will not easily be able to fit into visual
half-hour blocks.  (This has already been the case since Fall 2020 due to Covid
response, but the intent is that the Fall 2021 changes will be long-term.)

Move code that handles this in month view to a new function,
start_and_end_timestamp(), in functions.inc, and use it instead.  Move cmp3()
which supports it as well.  Add code to day and week views to determine the
midnight bounds of each day's display, which affects how the times for a
multi-day event are shown each day.