More robustly handle small gaps between events
Increase yellow "caution"/"starting soon" time from 15 minutes to 20 minutes,
as the time between classes considered consecutive is now often 15 or 20
minutes, and times are entered more precisely due to the new scheduling
procedures effective Fall 2021.

Additionally, don't count gaps of 30 minutes or less when considering whether a
room is in continuous use for the purposes of displaying the "busy until" time.